+++ client/stats/host.go
@@ -4,6 +4,7 @@ import (
 	"log"
 	"math"
 	"runtime"
+	"sync"
 	"time"
 
 	"github.com/shirou/gopsutil/cpu"
@@ -19,6 +20,7 @@ type HostStats struct {
 	Memory           *MemoryStats
 	CPU              []*CPUStats
 	DiskStats        []*DiskStats
+	AllocDirStats    *DiskStats
 	Uptime           uint64
 	Timestamp        int64
 	CPUTicksConsumed float64
@@ -52,32 +54,45 @@ type DiskStats struct {
 	InodesUsedPercent float64
 }
 
+// NodeStatsCollector is an interface which is used for the puproses of mocking
+// the HostStatsCollector in the tests
+type NodeStatsCollector interface {
+	Collect() error
+	Stats() *HostStats
+}
+
 // HostStatsCollector collects host resource usage stats
 type HostStatsCollector struct {
 	clkSpeed        float64
 	numCores        int
 	statsCalculator map[string]*HostCpuStatsCalculator
 	logger          *log.Logger
+	hostStats       *HostStats
+	hostStatsLock   sync.RWMutex
+	allocDir        string
 }
 
-// NewHostStatsCollector returns a HostStatsCollector
-func NewHostStatsCollector(logger *log.Logger) *HostStatsCollector {
+// NewHostStatsCollector returns a HostStatsCollector. The allocDir is passed in
+// so that we can present the disk related statistics for the mountpoint where
+// the allocation directory lives
+func NewHostStatsCollector(logger *log.Logger, allocDir string) *HostStatsCollector {
 	numCores := runtime.NumCPU()
 	statsCalculator := make(map[string]*HostCpuStatsCalculator)
 	collector := &HostStatsCollector{
 		statsCalculator: statsCalculator,
 		numCores:        numCores,
 		logger:          logger,
+		allocDir:        allocDir,
 	}
 	return collector
 }
 
 // Collect collects stats related to resource usage of a host
-func (h *HostStatsCollector) Collect() (*HostStats, error) {
+func (h *HostStatsCollector) Collect() error {
 	hs := &HostStats{Timestamp: time.Now().UTC().UnixNano()}
 	memStats, err := mem.VirtualMemory()
 	if err != nil {
-		return nil, err
+		return err
 	}
 	hs.Memory = &MemoryStats{
 		Total:     memStats.Total,
@@ -89,7 +104,7 @@ func (h *HostStatsCollector) Collect() (*HostStats, error) {
 	ticksConsumed := 0.0
 	cpuStats, err := cpu.Times(true)
 	if err != nil {
-		return nil, err
+		return err
 	}
 	cs := make([]*CPUStats, len(cpuStats))
 	for idx, cpuStat := range cpuStats {
@@ -113,7 +128,7 @@ func (h *HostStatsCollector) Collect() (*HostStats, error) {
 
 	partitions, err := disk.Partitions(false)
 	if err != nil {
-		return nil, err
+		return err
 	}
 	var diskStats []*DiskStats
 	for _, partition := range partitions {
@@ -122,32 +137,62 @@ func (h *HostStatsCollector) Collect() (*HostStats, error) {
 			h.logger.Printf("[WARN] client: error fetching host disk usage stats for %v: %v", partition.Mountpoint, err)
 			continue
 		}
-		ds := DiskStats{
-			Device:            partition.Device,
-			Mountpoint:        partition.Mountpoint,
-			Size:              usage.Total,
-			Used:              usage.Used,
-			Available:         usage.Free,
-			UsedPercent:       usage.UsedPercent,
-			InodesUsedPercent: usage.InodesUsedPercent,
-		}
-		if math.IsNaN(ds.UsedPercent) {
-			ds.UsedPercent = 0.0
-		}
-		if math.IsNaN(ds.InodesUsedPercent) {
-			ds.InodesUsedPercent = 0.0
-		}
-		diskStats = append(diskStats, &ds)
+		ds := h.toDiskStats(usage, &partition)
+		diskStats = append(diskStats, ds)
 	}
 	hs.DiskStats = diskStats
 
+	// Getting the disk stats for the allocation directory
+	usage, err := disk.Usage(h.allocDir)
+	if err != nil {
+		return err
+	}
+	hs.AllocDirStats = h.toDiskStats(usage, nil)
+
 	uptime, err := host.Uptime()
 	if err != nil {
-		return nil, err
+		return err
 	}
 	hs.Uptime = uptime
 
-	return hs, nil
+	h.hostStatsLock.Lock()
+	defer h.hostStatsLock.Unlock()
+	h.hostStats = hs
+	return nil
+}
+
+// Stats returns the host stats that has been collected
+func (h *HostStatsCollector) Stats() *HostStats {
+	h.hostStatsLock.RLock()
+	defer h.hostStatsLock.RUnlock()
+	return h.hostStats
+}
+
+// toDiskStats merges UsageStat and PartitionStat to create a DiskStat
+func (h *HostStatsCollector) toDiskStats(usage *disk.UsageStat, partitionStat *disk.PartitionStat) *DiskStats {
+	if usage == nil {
+		return nil
+	}
+	ds := DiskStats{
+		Size:              usage.Total,
+		Used:              usage.Used,
+		Available:         usage.Free,
+		UsedPercent:       usage.UsedPercent,
+		InodesUsedPercent: usage.InodesUsedPercent,
+	}
+	if math.IsNaN(ds.UsedPercent) {
+		ds.UsedPercent = 0.0
+	}
+	if math.IsNaN(ds.InodesUsedPercent) {
+		ds.InodesUsedPercent = 0.0
+	}
+
+	if partitionStat != nil {
+		ds.Device = partitionStat.Device
+		ds.Mountpoint = partitionStat.Mountpoint
+	}
+
+	return &ds
 }
 
 // HostCpuStatsCalculator calculates cpu usage percentages